build: fix envlist in tox.ini
The minimal Python version that I'm going to support
is Python v3.7, so drop all versions below it.

Also, add `basepython=...` options, because otherwise,
if needed python version isn't installed on the host system,
tox will silently replace it with a different (closest)
version of Python, which is definitely not what I want.