Revert "drop `SentinelValue.__new__`, and move its code to `sentinel()` fn"
This reverts commit 36c0b0dcccba0946532698daee73021d777ace38.

Because I discovered that creating objects via `sentinel()` function
doesn't work well with type annotations.

After a `sentinel()` object is created, it becomes impossible
to come up with a type annotation that matches to the sentinel.

PEP 661 suggest to use `Literal[SENTINEL]`, but I checked it,
and it doesn't work (at least with `mypy` on my machine).

So, I decided to revert the latest commit, andthus  make it possible
using the `SentinelValue` class alone (without the `sentinel()`
function), which allows to write code perfectly firendly
towards type checkers and most other kinds of static analysis.