change `__class__` of SentinelValue instance on the fly
This is needed to support hot code reloading featurs in IDEs,
where it sends parts of code to Python shell without full restart.

So, when that happens, code that subclasses SentinelValue may change,
so here I catch such change and dynamically update the `__class__`
attribute, setting it to the new class.

Hacky, but works.