fix: py.typed is installed to site-packages directory
In the previous commit, I included `py.typed` file
into the distribution.

That was a mistake, because `sentinel_value` is not a package,
but an individual module, that is installed directly to the
`site-packages` directory, and thus `py.typed` is installed nearby.

So, long story short: `py.typed` pollutes `site-packages`.

In order to avoid it, I had to convert `sentinel_value`
to a package.

The directory structure looks a bit redundant, but it works.